# hinted collect+parallel = ~400% faster!