--- a/R/zzz.R
+++ b/R/zzz.R
@@ -147,7 +147,7 @@ pl$mem_address = mem_address
 
 
   #see doc below, R CMD check did not like this function def
-  pl$select = pl$DataFrame(list())$select
+  pl$select = .pr$DataFrame$new()$select
 
   lockEnvironment(pl,bindings = TRUE)