--- a/inst/misc/simple_benchmark.R
+++ b/inst/misc/simple_benchmark.R
@@ -23,7 +23,12 @@ x = bench::mark(
   },
   df2 <- pl$DataFrame(nycflights13::flights),
   df3 <- pl$DataFrame(nyf),
-
+  df4 <- .pr$DataFrame$new_par_from_list(as.list(nycflights13::flights)),
+  df5 <- .pr$DataFrame$new_par_from_list(as.list(nyf)),
+  df_fa1 <- pl$from_arrow(arrow_table(nycflights13::flights)),
+  df_fa2 <- pl$from_arrow(arrow_table(nyf)),
+  df6 <- pl$DataFrame(nycflights13::flights,parallel = TRUE),
+  df7 <- pl$DataFrame(nyf,parallel = TRUE),
   rpolars_df = df$as_data_frame(),
   rpolars_list = df$to_list(),
   rpolars_list_no_char = df_no_char$to_list(),