--- a/man/pl.Rd
+++ b/man/pl.Rd
@@ -5,7 +5,7 @@
 \alias{pl}
 \title{The complete rpolars public API.}
 \format{
-An object of class \code{environment} of length 55.
+An object of class \code{environment} of length 56.
 }
 \usage{
 pl