rpolars is seemingly now quite faster than r-arrow to import from R vectors/data.frame.
rpolars can now threadsafe (emperically) read from R mem from multiple threads and thereby become ~4 times faster than r-arrow on a 4cpu computer. Without parallelism rpolars and r-arrow are very similar in speed.


##Before 75ms to read nycflights13::flights
```R
> print(x)
# A tibble: 11 × 13
   expression                                                       min   median `itr/sec`
   <bch:expr>                                                  <bch:tm> <bch:tm>     <dbl>
 1 nanoarrow                                                     25.2µs  31.95µs  24291.  
 2 nanoarrow_df                                                  8.32ms  12.13ms     60.1 
 3 arrow_table                                                   3.68ms   4.95ms    206.  
 4 df2 <- pl$DataFrame(nycflights13::flights)                   66.12ms  75.76ms     13.1 
 5 df3 <- pl$DataFrame(nyf)                                     44.42ms  46.35ms     20.8 
 6 df_fa1 <- pl$from_arrow(arrow_table(nycflights13::flights))  72.44ms  77.68ms     12.9 
 7 df_fa2 <- pl$from_arrow(arrow_table(nyf))                    32.35ms  34.09ms     26.5 
 8 rpolars_df                                                  108.86ms 111.28ms      8.93
 9 rpolars_list                                                104.43ms 116.15ms      8.82
10 rpolars_list_no_char                                         40.41ms   46.8ms     21.3 
11 rpolars_uwind                                               109.04ms 111.02ms      9.00
# … with 9 more variables: mem_alloc <bch:byt>, `gc/sec` <dbl>, n_itr <int>, n_gc <dbl>,
#   total_time <bch:tm>, result <list>, memory <list>, time <list>, gc <list>
# ℹ Use `colnames()` to see all variable names

```


##After 18ms to read nycflights13::flights
```R
1 nanoarrow                                                               24.05µs  34.42µs  23993.  
 2 nanoarrow_df                                                            15.29ms  18.17ms     53.3 
 3 arrow_table                                                              3.67ms   5.19ms    200.  
 4 df2 <- pl$DataFrame(nycflights13::flights)                              59.15ms  68.77ms     13.6 
 5 df3 <- pl$DataFrame(nyf)                                                 33.3ms  38.82ms     25.4 
 6 df4 <- .pr$DataFrame$new_par_from_list(as.list(nycflights13::flights))  15.44ms  17.01ms     46.7 
 7 df5 <- .pr$DataFrame$new_par_from_list(as.list(nyf))                     8.01ms  11.24ms     70.4 
 8 df_fa1 <- pl$from_arrow(arrow_table(nycflights13::flights))              75.4ms  98.49ms      9.76
 9 df_fa2 <- pl$from_arrow(arrow_table(nyf))                               34.47ms  36.34ms     25.6 
10 df6 <- pl$DataFrame(nycflights13::flights, parallel = TRUE)             18.18ms  22.41ms     42.5 
11 df7 <- pl$DataFrame(nyf, parallel = TRUE)                               12.46ms  14.79ms     43.3 
12 rpolars_df                                                             124.53ms 130.51ms      7.47
13 rpolars_list                                                           119.74ms 127.73ms      7.78
14 rpolars_list_no_char                                                    41.52ms  52.74ms     19.6 
15 rpolars_uwind                                                          114.38ms 126.45ms      7.77
```